+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## 1.2.0-dev
+## 1.1.4
+
+* Correctly paginate unconstrained queries
+
+## 1.1.3
+
+* Correctly handle nested joins
 
 ## 1.1.2